Unify run numbers to u32 across the public surface
The store layer exposed run numbers as `i32` (matching PostgreSQL's signed
`integer`) while the handler-facing `Context`/`JournalEntry` exposed `u32`, with
`.max(0) as u32` conversions bridging the two. `JobRecord::run_count`,
`JournalRecord::run_no`, `JournalAppend::run_no`, and the `run_no` argument of
`settle`/`recover`/`extend_lease` are now `u32`, so a run number is a single
type everywhere above the database.

The signedness is confined to the PostgreSQL adapter: it reads the `integer`
column into `u32` (always non-negative) and converts back through
`run_no_to_db`, which rejects a value above `i32::MAX` with the new
`Error::RunNumberOutOfRange` instead of letting `as i32` wrap to a negative
epoch. Such a value cannot arise from the adapter's own path; the guard covers a
hand-built `JournalAppend`.

+++ tests/notifications.rs
@@ -86,7 +86,7 @@ fn new_job(id: Ulid) -> NewJob {
     }
 }
 
-fn journal(run_no: i32, outcome: JournalOutcome) -> JournalAppend {
+fn journal(run_no: u32, outcome: JournalOutcome) -> JournalAppend {
     JournalAppend {
         kind: KIND.to_owned(),
         run_no,
@@ -103,7 +103,7 @@ async fn enqueue_and_claim(
     store: &impl Store,
     rx: &mut UnboundedReceiver<()>,
     lease: Duration,
-) -> (Ulid, i32) {
+) -> (Ulid, u32) {
     let id = Ulid::new();
     store.enqueue(&new_job(id)).await.expect("enqueue");
     let kinds = vec![KIND.to_owned()];
@@ -285,7 +285,7 @@ async fn recover_notifies() {
         .recover(
             id,
             Utc::now(),
-            run_no + 1,
+            (run_no + 1) as i32,
             run_no,
             journal(run_no, JournalOutcome::StaleRecovered),
         )
